added some more validations

+++ src/screens/DeletarHabilidade.js
@@ -9,8 +9,28 @@ class DeletarHabilidade extends React.Component{
     state = {
       id: '',
     }
+
+    validate = () =>{
+      const errors = [];
+
+      if(!this.state.nome){
+        errors.push('Campo id é obrigário');
+      }
+
+      return errors;
+    }
   
     delete = () => {
+
+        const errors = this.validate();
+
+        if(errors.length > 0){
+          errors.forEach((message, index) => {
+            showErrorMessage(message);
+          });
+          return false;
+        }
+
         axios.delete(`http://localhost:8080/api/habilidade/${this.state.id}`,
             {
